PYTHON/study-python/pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "stridentunlawfulsymbol"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.7"
bs4 = "^0.0.1"
request = "^2019.4"
[tool.poetry.dev-dependencies]
